# Backend/ApiGatewate/docs/swagger.yaml
basePath: /api/v1
definitions:
  authv1.CheckAccessTokenRequest:
    properties:
      access_token:
        type: string
    type: object
  authv1.CheckAccessTokenResponse:
    properties:
      user_id:
        type: integer
    type: object
  authv1.CheckServiceTokenRequest:
    properties:
      service_token:
        type: string
    type: object
  authv1.CheckServiceTokenResponse:
    properties:
      valid:
        type: boolean
    type: object
  authv1.GenerateAuthCodeRequest:
    properties:
      user_id:
        type: integer
    type: object
  authv1.GenerateAuthCodeResponse:
    properties:
      code:
        type: string
    type: object
  authv1.GenerateServiceTokenRequest:
    properties:
      service_name:
        type: string
    type: object
  authv1.GenerateServiceTokenResponse:
    properties:
      token:
        type: string
    type: object
  authv1.GetRoleRequest:
    properties:
      user_id:
        type: integer
    type: object
  authv1.GetRoleResponse:
    properties:
      role:
        type: string
    type: object
  authv1.LoginResponse:
    properties:
      access_token:
        type: string
      id:
        type: integer
    type: object
  authv1.LogoutResponse:
    properties:
      success:
        type: boolean
    type: object
  authv1.RegisterResponse:
    properties:
      success:
        type: boolean
    type: object
  authv1.SetRoleRequest:
    properties:
      role:
        type: string
      user_id:
        type: integer
    type: object
  authv1.SetRoleResponse:
    properties:
      success:
        type: boolean
    type: object
  authv1.VerifyRequest:
    properties:
      code:
        type: string
      user_id:
        type: integer
    type: object
  authv1.VerifyResponse:
    properties:
      verified:
        type: boolean
    type: object
  entities.FileResp:
    properties:
      filename:
        type: string
      url:
        type: string
    type: object
  entities.LoginRequest:
    properties:
      password:
        maxLength: 50
        minLength: 8
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  entities.LogoutRequest:
    properties:
      access_token:
        type: string
    required:
    - access_token
    type: object
  entities.RegisterRequest:
    properties:
      password:
        maxLength: 50
        minLength: 8
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  entities.UploadResponse:
    properties:
      files:
        items:
          $ref: '#/definitions/entities.FileResp'
        type: array
    type: object
host: cookhub.space
info:
  contact: {}
  description: API Gatewate
  title: API Gatewate
  version: "1.0"
paths:
  /auth/check_access_token:
    post:
      consumes:
      - application/json
      description: Check access token validity
      operationId: CheckAccessToken
      parameters:
      - description: request
        in: body
        name: request
        schema:
          $ref: '#/definitions/authv1.CheckAccessTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authv1.CheckAccessTokenResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
        "503":
          description: Service Unavailable
      summary: Check Access Token
      tags:
      - Auth
  /auth/check_service_token:
    post:
      consumes:
      - application/json
      description: Check service token validity
      operationId: CheckServiceToken
      parameters:
      - description: request
        in: body
        name: request
        schema:
          $ref: '#/definitions/authv1.CheckServiceTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authv1.CheckServiceTokenResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
        "503":
          description: Service Unavailable
      summary: Check Service Token
      tags:
      - Auth
  /auth/generate_auth_code:
    post:
      consumes:
      - application/json
      description: Generate authentication code
      operationId: GenerateAuthCode
      parameters:
      - description: request
        in: body
        name: request
        schema:
          $ref: '#/definitions/authv1.GenerateAuthCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authv1.GenerateAuthCodeResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
        "503":
          description: Service Unavailable
      summary: Generate Auth Code
      tags:
      - Auth
  /auth/generate_service_token:
    post:
      consumes:
      - application/json
      description: Generate service token
      operationId: GenerateServiceToken
      parameters:
      - description: request
        in: body
        name: request
        schema:
          $ref: '#/definitions/authv1.GenerateServiceTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authv1.GenerateServiceTokenResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
        "503":
          description: Service Unavailable
      summary: Generate Service Token
      tags:
      - Auth
  /auth/get_role:
    post:
      consumes:
      - application/json
      description: Get user role
      operationId: GetRole
      parameters:
      - description: request
        in: body
        name: request
        schema:
          $ref: '#/definitions/authv1.GetRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authv1.GetRoleResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
        "503":
          description: Service Unavailable
      summary: Get Role
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: Login
      parameters:
      - description: login
        in: body
        name: login
        schema:
          $ref: '#/definitions/entities.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authv1.LoginResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
        "503":
          description: Service Unavailable
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout
      operationId: Logout
      parameters:
      - description: logout
        in: body
        name: logout
        schema:
          $ref: '#/definitions/entities.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authv1.LogoutResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
        "503":
          description: Service Unavailable
      summary: Logout
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register
      operationId: Register
      parameters:
      - description: register
        in: body
        name: register
        schema:
          $ref: '#/definitions/entities.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authv1.RegisterResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
        "503":
          description: Service Unavailable
      summary: Register
      tags:
      - Auth
  /auth/set_role:
    post:
      consumes:
      - application/json
      description: Set user role
      operationId: SetRole
      parameters:
      - description: request
        in: body
        name: request
        schema:
          $ref: '#/definitions/authv1.SetRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authv1.SetRoleResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
        "503":
          description: Service Unavailable
      summary: Set Role
      tags:
      - Auth
  /auth/verify:
    post:
      consumes:
      - application/json
      description: Verify authentication
      operationId: Verify
      parameters:
      - description: request
        in: body
        name: request
        schema:
          $ref: '#/definitions/authv1.VerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authv1.VerifyResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
        "503":
          description: Service Unavailable
      summary: Verify
      tags:
      - Auth
  /media/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload media
      operationId: Upload media
      parameters:
      - collectionFormat: csv
        description: files
        in: formData
        items:
          type: file
        name: files
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.UploadResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
        "503":
          description: Service Unavailable
      summary: Upload media
      tags:
      - media
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"